Added package folder. Added Variable Mapping Names to Database. Added better error messages and error sates to gb2rpg and cppRPGEnv. Added clean command to gb2rpg.

+++ src/tools/gb2rpg/src/Map.cpp
@@ -152,7 +152,8 @@ std::vector<lcf::EventCommand> Map::setupMapRomHeader(int numLabels)
     // ByteOffset = (ByteOffset / BYTES_PER_VAR):
     mapRomHeader.push_back(lcf::EventCommandFactory::GenControlVariable(
         lcf::ControlVariable::Type::SINGLE_VARIABLE,
-        VARMAPPING::BYTE_OFFSET_ID, VARMAPPING::BYTE_OFFSET_ID,
+        static_cast<int>(VARMAPPING::BYTE_OFFSET_ID), 
+        static_cast<int>(VARMAPPING::BYTE_OFFSET_ID),
         lcf::ControlVariable::Operation::OPERATION_DIV,
         lcf::ControlVariable::OperandTypes::OPERANDS_CONSTANT,
         MEMORYSIZES::BYTES_PER_VAR, 0));
@@ -160,15 +161,17 @@ std::vector<lcf::EventCommand> Map::setupMapRomHeader(int numLabels)
     // LabelID = ByteOffset
     mapRomHeader.push_back(lcf::EventCommandFactory::GenControlVariable(
         lcf::ControlVariable::Type::SINGLE_VARIABLE,
-        VARMAPPING::LABEL_ID, VARMAPPING::LABEL_ID,
+        static_cast<int>(VARMAPPING::LABEL_ID), 
+        static_cast<int>(VARMAPPING::LABEL_ID),
         lcf::ControlVariable::Operation::OPERATION_SET,
         lcf::ControlVariable::OperandTypes::OPERANDS_VARIABLE,
-        VARMAPPING::BYTE_OFFSET_ID, 0));
+        static_cast<int>(VARMAPPING::BYTE_OFFSET_ID), 0));
 
     // LabelID += 1
     mapRomHeader.push_back(lcf::EventCommandFactory::GenControlVariable(
         lcf::ControlVariable::Type::SINGLE_VARIABLE,
-        VARMAPPING::LABEL_ID, VARMAPPING::LABEL_ID,
+        static_cast<int>(VARMAPPING::LABEL_ID), 
+        static_cast<int>(VARMAPPING::LABEL_ID),
         lcf::ControlVariable::Operation::OPERATION_ADD,
         lcf::ControlVariable::OperandTypes::OPERANDS_CONSTANT,
         1, 0));
@@ -193,7 +196,7 @@ std::vector<lcf::EventCommand> Map::setupMapRomLabel(int labelID, int numLabels,
     // IF(LabelID < X)
     mapRomLabel.push_back(lcf::EventCommandFactory::GenConditionalBranch(
         lcf::ConditionalBranch::Type::VARIABLE,
-        VARMAPPING::LABEL_ID, 
+        static_cast<int>(VARMAPPING::LABEL_ID), 
         lcf::ConditionalBranch::RHSType::USE_CONSTANT, labelID, 
         lcf::ConditionalBranch::Comparison::LESS, false));
     
@@ -219,7 +222,7 @@ std::vector<lcf::EventCommand> Map::setupMapRomLabel(int labelID, int numLabels,
     // IF(LabelID > X)
     mapRomLabel.push_back(lcf::EventCommandFactory::GenConditionalBranch(
         lcf::ConditionalBranch::Type::VARIABLE,
-        VARMAPPING::LABEL_ID, 
+        static_cast<int>(VARMAPPING::LABEL_ID), 
         lcf::ConditionalBranch::RHSType::USE_CONSTANT, labelID, 
         lcf::ConditionalBranch::Comparison::GREATER, false));
 
@@ -244,7 +247,8 @@ std::vector<lcf::EventCommand> Map::setupMapRomLabel(int labelID, int numLabels,
     // READVAR1 = LABELXVALUE1
     mapRomLabel.push_back(lcf::EventCommandFactory::GenControlVariable(
         lcf::ControlVariable::Type::SINGLE_VARIABLE,
-        VARMAPPING::READ_VAR_1, VARMAPPING::READ_VAR_1,
+        static_cast<int>(VARMAPPING::READ_VAR_1), 
+        static_cast<int>(VARMAPPING::READ_VAR_1),
         lcf::ControlVariable::Operation::OPERATION_SET,
         lcf::ControlVariable::OperandTypes::OPERANDS_CONSTANT,
         firstVar, 0));
@@ -252,7 +256,8 @@ std::vector<lcf::EventCommand> Map::setupMapRomLabel(int labelID, int numLabels,
     // READVAR2 = LABELXVALUE2
     mapRomLabel.push_back(lcf::EventCommandFactory::GenControlVariable(
         lcf::ControlVariable::Type::SINGLE_VARIABLE,
-        VARMAPPING::READ_VAR_2, VARMAPPING::READ_VAR_2,
+        static_cast<int>(VARMAPPING::READ_VAR_2), 
+        static_cast<int>(VARMAPPING::READ_VAR_2),
         lcf::ControlVariable::Operation::OPERATION_SET,
         lcf::ControlVariable::OperandTypes::OPERANDS_CONSTANT,
         secondVar, 0));
